kymotracker: move kymoline tests into dedicated test file
